requirements.txt
# ===============================
# Frontend
# ===============================
streamlit>=1.31.0

# ===============================
# Agent & Graph Framework
# ===============================
langgraph>=0.0.40
langchain>=0.1.16
langchain-core>=0.1.45

# ===============================
# LLM (Gemini)
# ===============================
google-generativeai
# ===============================
# Configuration & Environment
# ===============================
python-dotenv>=1.0.1

# ===============================
# Data Validation
# ===============================
pydantic>=2.6.0
typing-extensions>=4.9.0

# ===============================
# Logging & Utilities
# ===============================
loguru>=0.7.2
